Extract workspace output formatting
